Lots of Arsenal things. See description.
Added Draedon weapon chargers and generators. As of pushing this, it uses dirt blocks as fuel. This will be changed when a sprite and name are decided for the actual fuel. These have been roughly tested in Multiplayer. No remaining bugs exist to my knowledge.

Added a schematic worldgen system. There are two .cs files that they use with autogenerated code. Please excuse their poor quality.

Added a rough base for pop GUIs (for book-reading and the like. A draedons logbook exists that uses this)

Adjusted Snake eyes to use a little less dust.

Note: This is NOT ready to be merged into master yet. A lot of placeholders are used.