Reworked Ethereal Subjugator
A variable was added to superhoming to adjust how much predictiveness is incorporated into the vector. this is so that the projectile doesn't go funky mode and copy all npc movement when extra updates are in the equation. plus, other people can use it if they want to for other situations. by default, it is 1f which keeps the predictiveness at normal behaviour. all other instances of superhoming still work as they did before the added variable, with no changes to their code.