Reworked Fallen Paladin's Hammer
FPH is now thrown like Pwnagehammer and after hitting the enemy will home back in on the player. It no longer explodes on hit, but after three successful hits a phantom hammer will shoot out of the player and fly at the last enemy hit, dealing huge damage and massive AoE.

Also adjusted Pwnagehammer's "gong" to be a bit louder.