Attempted to fix abyss' chests unlocking in multiplayer creating corrupted tiles that cause crashes.
I added a multiplayer check; should theoretically ensure only runs the unlock method once.

Changed the code to set AbleToUnlockChests to false once the unlock method is over to prevent chests being generated unlocked. Updated AbyssTreasureChest's comment to reflect that.

The downed Skeletron bool check will be left as a failsave as the failsave has been extremely helpful for players in the bugs forum as they can just set skeletron to dead and avoid the method and still be able to unlock chests and continue their playthroughs, but asking players to go install Toasty's QoL mod or CalTestHelpers is not ideal