Rework from Yatagarasu (patreon donor for Cosmic Shiv), which now shoots a homing star trail to a target, spawning an invisible, non-damaging aura that sticks to the target, continously spawning extremely fast blades for 5 seconds. I didn't add a limit for how many auras can exist at once, but I added limits to how many can be applied at once. Galaxy Star is no longer a thing, but the texture is used for Cosmic Shiv Trail now, so I renamed the file to match.
My intention with this weapon is to be a slow firing projectile-focused weapon that can stack up a bit on one target, and be able to keep up with Yharon, hence the accelerating homing and high speed blades. Feel free to nerf the penetration of the blades if they're too op against Thanatos.

This one was a big step-up from my previous projects (which was basically just making very basic weapons or slightly modifying ExampleMod code, and I'm overall extremely happy with how it went. Code might be pretty bad since I only had 2 weeks of modding experience by the time I finished this, so the code might need some cleaning up, and if the draws are too laggy, feel free to nerf the values a bit (although I tweaked with them a lot to get them looking the best)